Refactor hooks
