+++ src/api/index.ts
@@ -1,7 +1,4 @@
 export * from "./getAvatarInventory";
 export * from "./getRawState";
 export * from "./stageTransaction";
-export * from "./useAvatarStates";
-export * from "./useGoldBalance";
-export * from "./useNextTxNonce";
 export * from "./shortcuts";
