+++ src/hooks/useAccountContext.ts
@@ -1,8 +1,6 @@
 import { ContextType, useContext } from "react";
 import AccountContext from "../contexts/Account";
 
-export default function useAccountContext(): ContextType<
-  typeof AccountContext
-> {
+export function useAccountContext(): ContextType<typeof AccountContext> {
   return useContext(AccountContext);
 }
