+++ src/hooks/useAvatarStates.ts
